split backend server into mongo handlers, google handlers, and left the routing in the original server file
